requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
requests==2.32.3
pydantic==2.9.2
python-multipart==0.0.9
pyyaml==6.0.2

numpy==1.26.4
pandas==2.2.3
opencv-python==4.10.0.84
Pillow==10.4.0

torch
torchaudio
torchvision

transformers==4.44.2
sentence-transformers==3.0.1
chromadb==0.5.5
openai-whisper==20240930

ffmpeg-python==0.2.0
pytest==8.3.3